+++ app/src/main/java/cityfreqs/com/pilfershushjammer/assist/AssistJammerService.java
@@ -1,0 +1,7 @@
+package cityfreqs.com.pilfershushjammer.assist;
+
+import android.service.voice.VoiceInteractionService;
+
+public class AssistJammerService extends VoiceInteractionService {
+    //
+}