+++ app/src/main/java/cityfreqs/com/pilfershushjammer/assist/AssistJammerSessionService.java
@@ -1,0 +1,12 @@
+package cityfreqs.com.pilfershushjammer.assist;
+
+import android.os.Bundle;
+import android.service.voice.VoiceInteractionSession;
+import android.service.voice.VoiceInteractionSessionService;
+
+public class AssistJammerSessionService extends VoiceInteractionSessionService {
+    @Override
+    public VoiceInteractionSession onNewSession(Bundle args) {
+        return new AssistJammerSession(this);
+    }
+}