+++ app/src/main/java/cityfreqs/com/pilfershushjammer/ui/SectionsPagerAdapter.java
@@ -105,4 +105,12 @@ public class SectionsPagerAdapter extends FragmentPagerAdapter {
     public void permissionCheckPassed(boolean permissionCheck) {
         audioBundle.putBoolean(AudioSettings.AUDIO_BUNDLE_KEYS[16], permissionCheck);
     }
+
+    public void setAssistBoolean(boolean setBoolean) {
+        audioBundle.putBoolean(AudioSettings.AUDIO_BUNDLE_KEYS[20], setBoolean);
+    }
+
+    public Bundle getAudioBundle() {
+        return audioBundle;
+    }
 }